1318 - move one-time-donation folder to client
